0x0F-python-object_relational_mapping

+++ 0x0F-python-object_relational_mapping/13-model_state_delete_a.py
@@ -1,32 +1,22 @@
 #!/usr/bin/python3
-"""
-Script that deletes all State objects with a name containing
-the letter 'a' from the database.
-"""
+# Deletes all State objects with a name containing
+# Script that deletes all State objects with a name containing the letter a.
+# Usage: ./13-model_state_delete_a.py <mysql username> /
+#                                     <mysql password> /
+#                                     <database name>
 import sys
 from sqlalchemy import create_engine
 from sqlalchemy.orm import sessionmaker
 from model_state import State
 
 if __name__ == "__main__":
-    # Create database engine
-    engine = create_engine('mysql+mysqldb://{}:{}@localhost:3306/{}'
+    engine = create_engine("mysql+mysqldb://{}:{}@localhost/{}"
                            .format(sys.argv[1], sys.argv[2], sys.argv[3]),
                            pool_pre_ping=True)
-
-    # Create a configured "Session" class
     Session = sessionmaker(bind=engine)
-
-    # Create a session
     session = Session()
 
-    # Query and delete all states with a name containing the letter 'a'
-    states_to_delete = session.query(State).filter(State.name.like('%a%')).all()
-    for state in states_to_delete:
-        session.delete(state)
-
-    # Commit changes
+    for state in session.query(State):
+        if "a" in state.name:
+            session.delete(state)
     session.commit()
-
-    # Close the session
-    session.close()
